Sagar_Dataharvesting/Real-Time DB/requirements.txt
aiomysql==0.2.0
dbutils==3.1.0
Flask==3.0.3
flask_socketio==5.3.6
mysql_connector_repackaged==0.3.1
pandas==2.2.2
PyMySQL==1.1.1
python-socketio==5.11.3
Requests==2.32.3
retrying==1.3.4
SQLAlchemy==2.0.25
urllib3==1.26.18
./dist/NorenRestApi-0.0.30-py2.py3-none-any.whl

